requirements.txt
google_api_python_client==2.154.0
google_auth_oauthlib==1.2.1
litellm==1.53.1
loguru==0.7.2
protobuf==5.29.0
pydantic==2.10.2
python-dotenv==1.0.1
python-telegram-bot==21.8
pytz==2024.2
sentry_sdk==2.19.0
sentry-sdk[flask]==2.19.0
tzlocal==5.2
flask~=3.1.0
flask[async]==3.1.0
sentry-sdk~=2.19.0
google-auth-oauthlib~=1.2.1
google-api-python-client~=2.154.0
uvicorn~=0.32.1
asgiref~=3.8.1
google-cloud-firestore~=2.19.0
aioboto3~=13.2.0
tinydb~=4.8.2